+++ crates/compiler/src/driver_constants.rs
@@ -98,7 +98,7 @@ pub mod addresses {
     );
 
     // MUST match `audio-driver/src/common_memmap.wiz`
-    pub const COMMON_DATA: u16 = 0x1000 - 28;
+    pub const COMMON_DATA: u16 = 0x1100 - 28;
 
     const _: () = assert!(
         _symbols::_LAST_LOADER_SYMBOL < DRIVER_CODE,